# Speed up conflict resolution.