… CHANGES.rst
… src/relstorage/tests/util.py