- Get committed data in bulk;
- Store replacement data in bulk.

This eliminates lots of database queries if there's more than one conflict. Since this happens while rows are locked, the faster this is, and the fewer round trips, the better. Results are 40-90%
improvements, depending on cache configuration. It does take a little more memory to store the committed data.

(The benchmark has half of the workers conflict on half of the objects.)

| Benchmark                                   | Master | This branch |
|---------------------------------------------|-----------------------------------------|-----------------------------------------|
| psycopg2_hf: update 100 conflicting objects (no cache) | 1.55 sec                                | 878 ms: 1.76x faster (-43%)             |
| psycopg2_hf: update 100 conflicting objects (cached) | 789 ms                                | 94.1 ms: 8.39x faster (-88%)          |
| mysqlclient_hf: update 100 conflicting objects (no cache) | 1.47 sec                                | 855 ms: 1.72x faster (-42%)             |
| mysqlclient_hf: update 100 conflicting objects (cached) | 657 ms                                | 122 ms: 5.39x faster (-81%)           |